fix: 修复代码审查发现的全部 Critical/Important 问题
- C1: 清除所有 XML mapper 中残留的 is_deleted 引用，统一为 del_flag = '0'
- C2: PrTransactionRecordSub 补充 item_group/item_code/houseId 字段
- C3: 新增 pr_house_log/pr_transaction_detail/pr_import_account DDL
- C4: PrHouseLog 改为继承 BaseEntity，使用 del_flag 约定
- I1: WechatRepairController 移除 Long.valueOf(UUID) 崩溃调用
- I2: generateRepairNo() 移至接口，Controller 不再直接引用 ServiceImpl
- I3: pr_inspection_plan start_time/end_time 从 time 改为 datetime
- I4: PrRepairPerson 补充 orgId 字段，selectByOrgId 添加过滤条件

+++ sdkj-modules/sdkj-thermal/src/main/java/org/sdkj/thermal/domain/bo/PrHeatArchiveBo.java
@@ -136,7 +136,7 @@ public class PrHeatArchiveBo extends BaseEntity {
     private Integer isStop;
 
     /** 起始读数 */
-    private BigDecimal startReading;
+    private Integer startReading;
 
     /** 总热量 */
     private BigDecimal totalHeat;
